+++ mod/racing_plus/BossRush.lua
@@ -90,7 +90,7 @@ function BossRush:PostUpdate()
 end
 
 function BossRush:CheckStart()
-  if not g.p:IsItemQueueEmpty() and
+  if g.run.touchedPickup and
      not g.run.bossRush.started and
      not g.run.bossRush.finished then
 
@@ -114,8 +114,9 @@ function BossRush:Start()
   -- Spawn a room clear delay NPC as a helper to keep the doors closed
   -- (otherwise, the doors will re-open on every frame)
   local roomClearDelayNPC = Isaac.Spawn(EntityType.ENTITY_ROOM_CLEAR_DELAY_NPC, 0, 0,
-                                        g:GridToPos(0, 0), g.zeroVector, nil)
+                                        g.zeroVector, g.zeroVector, nil)
   roomClearDelayNPC:ClearEntityFlags(EntityFlag.FLAG_APPEAR) -- 1 << 2
+  roomClearDelayNPC.EntityCollisionClass = EntityCollisionClass.ENTCOLL_NONE -- 0
   Isaac.DebugString("Spawned the \"Room Clear Delay NPC\" custom entity (for the Boss Rush).")
 
   -- Close the door