+++ mod/racing_plus/FastTravel.lua
@@ -428,7 +428,8 @@ function FastTravel:CheckTrapdoor()
     -- Stage 3 is actiated when the screen is black
     g.run.trapdoor.state = FastTravel.state.SCREEN_IS_BLACK
     g.run.trapdoor.floor = stage
-    Sprites:Init("black", "black")
+    Sprites:Init("black", "black") -- Cover the screen with a big black sprite
+    MinimapAPI.Config.Disable = true -- We must also disable the custom minimap when we want the screen to be black
     FastTravel:GotoNextFloor(g.run.trapdoor.upwards)
 
   elseif g.run.trapdoor.state == FastTravel.state.POST_NEW_ROOM_2 and
@@ -521,6 +522,7 @@ function FastTravel:CheckNewFloor()
 
     -- Remove the black sprite to reveal the new floor
     Sprites:Init("black", 0)
+    MinimapAPI.Config.Disable = false
 
     local pos = g.r:GetCenterPos()
     if g.g.Difficulty >= Difficulty.DIFFICULTY_GREED and -- 2