+++ mod/racing_plus/PostCurseEval.lua
@@ -3,7 +3,6 @@ local PostCurseEval = {}
 -- ModCallbacks.MC_POST_CURSE_EVAL (12)
 function PostCurseEval:Main(curses)
   -- Disable all curses
-  Isaac.DebugString("Disabling curse: " .. tostring(curses))
   return 0
 end